+++ ricaun.RevitTest.Tests/Tests.cs
@@ -1,0 +1,38 @@
+﻿using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.UI;
+using NUnit.Framework;
+using System;
+
+namespace ricaun.RevitTest.Tests
+{
+    public class Tests
+    {
+        [Test]
+        public void RevitTests_VersionName(UIApplication uiapp)
+        {
+            Assert.IsNotNull(uiapp);
+            Console.WriteLine(uiapp.Application.VersionName);
+        }
+
+        [Test]
+        public void RevitTests_VersionName(Application application)
+        {
+            Assert.IsNotNull(application);
+            Console.WriteLine(application.VersionName);
+        }
+
+        [Test]
+        public void RevitTests_VersionName(UIControlledApplication uiControlledApplication)
+        {
+            Assert.IsNotNull(uiControlledApplication);
+            Console.WriteLine(uiControlledApplication.ControlledApplication.VersionName);
+        }
+
+        [Test]
+        public void RevitTests_VersionName(ControlledApplication controlledApplication)
+        {
+            Assert.IsNotNull(controlledApplication);
+            Console.WriteLine(controlledApplication.VersionName);
+        }
+    }
+}